Build modules from source by default
Most branches, most of the time, will build modules from source. By
changing the default we make it easier to switch a whole branch to use
module prebuilts by requiring only one CL to be committed on that branch
and nowhere else.

Before this change, dozens of Soong targets needed to be updated when
deciding to start using prebuilts for modules and sdk stubs. In
addition, the typical need is to enable prebuilts on a single branch at
the time, which resulted in a dozens of "D0 N0T MERGE ANYWHERE" CLs,
causing a lot of confusion.

Bug: 184842499
Test: (on a branch with ModuleMetadataGoogle enabled as prebuilt):
  m out/target/product/sunfish/product/app/ModuleMetadataGoogle/ModuleMetadataGoogle.apk
  (verified with `aapt2 dump badging` that apk is built from source)
  MODULE_BUILD_FROM_SOURCE=false m out/target/product/sunfish/product/app/ModuleMetadataGoogle/ModuleMetadataGoogle.apk
  (verified with `aapt2 dump badging` that apk is prebuilt)

Change-Id: I625c48640c4bb1edf360605a1daac293a4a0a07a
Signed-off-by: Jason Edson <jaysonedson@gmail.com>